requirements.txt
pandas
sklearn
numpy
keras
python-dotenv
requests
tensorflow